// src/core/model/banner.ts
export interface Banner {
	createDate?: Date;
	id?: number;
	name?: string;
	subTitle?: string;
	shortDescription?: string;
	isActive?: boolean;
	image?: string;
	description?: string;
	url?: string;
	updateDate?: Date;
	deleteDate?: Date;
	createById?: number;
	updateById?: number;
	deleteById?: number;
	isDelete?: boolean;
}